requirements-web.txt
# Web 服务依赖 (v3.0)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-socketio==5.10.0
aiofiles==23.2.1

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# 数据处理
pandas==2.1.3
numpy==1.26.2

# 监控和指标
psutil==5.9.6

# 其他
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
